chore: Extract columnSizes.
